Support viewing/editing secrets with vaultpath
If a secret has a vault path, show the reference to the URL in Vault
instead of the value of the secret.

For existing secrets/resources, the UI will "detect" whether
the secret is fetched directly from Fasit's database or via Vault
(with the Vault path). It renders the input dialog accordingly.

With resourceType, we can specify which style of secret we
prefer to use for newly created resources.